#13 Image on fullscreen never changes (Android) https://github.com/keensoft/FullScreenImage-Cordova-Plugin/issues/13
Fix solution https://github.com/L-VE:
I had the same problem opening a base64 image string. The file that was opening in fullscreen, was written in storage under the name of output.jpg. Then I took a look at the java code and noticed that all the data of the file I was trying to open was just appended to the already existing "output.jpg" file. That's because of this line of code:

#13;